feat: update shifts

+++ api-hooks/shift-templates/use-remove-shift-template.ts
@@ -1,8 +1,10 @@
 import {useMutation, useQueryClient} from '@tanstack/react-query';
 import {api} from '@/lib/axios';
+import {useOwnerId} from "@/hooks/use-owner-id";
 
 
 export const useRemoveShiftTemplate = () => {
+    const ownerId = useOwnerId();
     const queryClient = useQueryClient();
 
     return useMutation<
@@ -11,7 +13,7 @@ export const useRemoveShiftTemplate = () => {
         string
     >({
         mutationFn: async (id: string) => {
-            await api.delete(`/api/shift-templates/${id}`);
+            await api.delete(`/api/shift-templates/${id}`, {params: {ownerId}});
         },
         onSuccess: () => {
             queryClient.invalidateQueries({queryKey: ['shift-templates']});
